Move bundle classes into a new package
Move them from `models` to `fhir`, which seems more appropriate.
