+++ api-application/src/main/java/gov/usds/vaccineschedule/api/pagination/AbstractPaginatingAndValidatingProvider.java
@@ -11,9 +11,9 @@ import ca.uhn.fhir.rest.server.exceptions.UnprocessableEntityException;
 import ca.uhn.fhir.validation.FhirValidator;
 import ca.uhn.fhir.validation.ValidationOptions;
 import ca.uhn.fhir.validation.ValidationResult;
+import gov.usds.vaccineschedule.api.fhir.BundleFactory;
+import gov.usds.vaccineschedule.api.fhir.OffsetLinkBuilder;
 import gov.usds.vaccineschedule.api.helpers.BaseURLProvider;
-import gov.usds.vaccineschedule.api.models.BundleFactory;
-import gov.usds.vaccineschedule.api.models.OffsetLinkBuilder;
 import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.InstantType;
